Send emails to applicants
