+++ package.json
@@ -54,7 +54,7 @@
     "js-yaml": "^3.4.3",
     "lodash": "^3.10.1",
     "mailchimp-api": "^2.0.7",
-    "mailgen": "^2.0.1",
+    "mailgen": "2.0.1",
     "markdown-it": "^5.1.0",
     "markdown-it-attrs": "^0.1.8",
     "moment": "^2.11.1",
